Bot API 5.1

+++ src/main/java/io/github/dhina17/tgbot/DhinaBot.java
@@ -37,7 +37,7 @@ public class DhinaBot extends AbilityBot {
 	}
 
 	@Override
-  	public int creatorId() {
+  	public long creatorId() {
     	return BotConfig.CREATOR_ID;
   	}
